Implements a Length struct
This implementation only supports meters, the only internal unit of measurement available. The current plan is to add a 'Unit' class and provide utility methods to interact with Length in a more flexible way. This will require further changes to the Length struct, possibly breaking changes, so at this point it is not ready for production.